Ordnung schaffen

+++ Meins/Blatt 1/Aufgabe2.py
@@ -1,8 +1,8 @@
 import cmath
 
-a = int(input("a: "))
-b = int(input("b: "))
-c = int(input("c: "))
+a = float(input("a: "))
+b = float(input("b: "))
+c = float(input("c: "))
 
 print("x1= " , (-b + cmath.sqrt(b**2 - 4*a*c))/(2*a))
 print("x2= " , (-b - cmath.sqrt(b**2 - 4*a*c))/(2*a))
